modules for news article creted
