+++ traits/src/sumarizer.rs
@@ -1,26 +1,12 @@
 
+pub mod summary;
+pub mod news_article;
 
-pub trait Summary {
-    fn summarize(&self) -> String;
-}
-
-pub struct NewsArticle {
-    pub name: String,
-    pub content: String,
-    pub author: String,
-    pub date: String,
-}
-
-impl Summary for NewsArticle {
-    fn summarize(&self) -> String {
-        format!("{} published on {}, \n {} \n by {}", self.name, self.date, self.content, self.author)
-    }
-}
-
-
+use crate::sumarizer::summary::Summary;
+// use crate::sumarizer::news_article;
 
 pub fn main() {
-    let news_article = NewsArticle{
+    let news_article = news_article::NewsArticle{
         name: String::from("Modiji"),
         content: String::from(
             "Paytm downfall for using nationalism and to sell more."
