+++ traits/src/sumarizer/summary.rs
@@ -1,0 +1,4 @@
+pub trait Summary {
+    fn summarize(&self) -> String;
+}
+
